refactor: useTVCOverTime to use react-query
