+++ src/hooks/api/urls.ts
@@ -2,7 +2,7 @@ export const getFundraisePosition = (account: string) => `positions/${account}/f
 export const getUnlockOverTime = () => "stats/unlock"
 export const getAPYOverTime = () => "stats/apy"
 export const getTVLOverTime = () => "stats_tvl"
-export const getTVCOverTIme = () => "stats_tvc"
+export const getTVCOverTime = () => "stats_tvc"
 export const getStakePositions = (account?: string) => (account ? `staking/${account}` : "staking")
 export const getCoveredProtocols = () => "protocols"
 export const getWaitForBlock = (blockNumber: number) => `wait-for-block?block=${blockNumber}`
